+++ pokewatcher/main/agent_state.c
@@ -25,6 +25,7 @@ static const char *STATE_STRINGS[] = {
     [PW_STATE_SLEEPING]  = "sleeping",
     [PW_STATE_REPORTING] = "reporting",
     [PW_STATE_DOWN]      = "down",
+    [PW_STATE_WAKEUP]    = "wakeup",
 };
 
 void pw_agent_state_init(void)